lacrei_saude_api/pyproject.toml
[tool.poetry]
name = "lacrei-saude-api"
version = "0.1.0"
description = ""
authors = ["lucasoliveira <blesslucas89@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-django = "^4.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
